+++ src/medias/constants.ts
@@ -1,0 +1,6 @@
+export const COM_DOMAIN = process.env.COM_DOMAIN ?? 'art';
+export const COM_URI = process.env.COM_URI + '.' + COM_DOMAIN;
+export const COM_API_URI = COM_URI + '/api';
+
+export const COM_TOKEN = process.env.COM_TOKEN ?? null;
+export const COM_IDENTITY = process.env.COM_IDENTITY ?? null;