+++ src/medias/models/domain/media-image.ts
@@ -1,0 +1,4 @@
+export interface MediaImage {
+  comid_id: number;
+  url: string;
+}