+++ src/medias/models/response/chapter-details.ts
@@ -1,0 +1,10 @@
+import type {ChapterDto} from '../dto/chapter.dto';
+
+export interface ChapterDetailsResponse {
+  chapter: ChapterDto;
+  dupGroupChapters: {
+    id: number;
+    hid: string;
+    groupName: string;
+  }[];
+}