+++ src/medias/models/response/chapter-list.ts
@@ -1,0 +1,6 @@
+import type {ChapterDto} from '../dto/chapter.dto';
+
+export interface ChapterListResponse {
+  last_page: number;
+  data: ChapterDto[];
+}