+++ src/medias/models/response/follows.ts
@@ -1,0 +1,6 @@
+import type {MediaDto} from '../dto/media.dto';
+
+export interface FollowsResponse {
+  last_page: number;
+  data: MediaDto[];
+}