+++ src/medias/models/shared/app-error.ts
@@ -1,0 +1,4 @@
+export interface AppError {
+  error: string;
+  status: number;
+}