+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## [1.1.0] - 2026-01-03
+
+### Changed
+
+- Migration from Javascript to Typescript
+- Migration to module export mode
+- Typed data, function, etc.
+- Improved architecture
+
+### Fixed
+
+- Refresh follows didn't send any media at first
+- Refresh specific comic didn't send the details at first
+- Getting the chapter images didn't send any data at first
+
 ## [1.0.1] - 2026-01-02
 
 ### Fixed